test: add test for the quiz topic selection screen
Changed the name of the QuizKtTest.kt file and class to get a direct link in Android Studio with the class.
Added new tests to the screen.
Reformated the Quiz file to the proper format.
Fixed a deprecated use of `androidx.compose.material.icons.filled.ArrowBack` to `androidx.compose.material.icons.automirrored.filled.ArrowBack`
